refactor(cli): extract file discovery
